added settings for family collector
